backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
httpx>=0.25.0
starlette>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# MCP Server Implementation
mcp>=1.0.0

# OpenAI Integration (Latest stable)
openai>=1.30.0
requests>=2.31.0
typing-extensions>=4.8.0